change determineIfCanShoot method to be divided under subclasses, not in the parent class.
