+++ src/main/java/nl/yer/middlemen/dreampoint/character/Enemy.java
@@ -1,26 +1,12 @@
 package nl.yer.middlemen.dreampoint.character;
 
-import static nl.yer.middlemen.dreampoint.game.Game.hiScore;
-
 public abstract class Enemy extends Character {
 
     private long id;
     private int addToHiScore;
 
     @Override
-    public boolean determineIfCanShoot() {
-        if (this.isCanDamage()) {
-            if (this.getCanTakeAmountOfShots() > 1) {
-                this.setCanTakeAmountOfShots(this.getCanTakeAmountOfShots() - 1);
-                return false;
-            } else {
-                hiScore += this.addToHiScore;
-                return true;
-            }
-        } else {
-            return false;
-        }
-    }
+    public abstract boolean determineIfCanShoot();
 
     @Override
     public long getId() {
