+++ src/main/java/nl/yer/middlemen/dreampoint/character/SmallEnemy.java
@@ -1,5 +1,7 @@
 package nl.yer.middlemen.dreampoint.character;
 
+import static nl.yer.middlemen.dreampoint.game.Game.hiScore;
+
 public class SmallEnemy extends Enemy {
 
     private long id;
@@ -8,6 +10,21 @@ public class SmallEnemy extends Enemy {
         this.setCanDamage(true);
         this.setCanTakeAmountOfShots(1);
         this.setAddToHiScore(10);
+    }
+
+    @Override
+    public boolean determineIfCanShoot() {
+        if (this.isCanDamage()) {
+            if (this.getCanTakeAmountOfShots() > 1) {
+                this.setCanTakeAmountOfShots(this.getCanTakeAmountOfShots() - 1);
+                return false;
+            } else {
+                hiScore += this.getAddToHiScore();
+                return true;
+            }
+        } else {
+            return false;
+        }
     }
 
     @Override
