+++ src/main/java/nl/yer/middlemen/dreampoint/playingfield/PlayingField.java
@@ -3,11 +3,9 @@ import nl.yer.middlemen.dreampoint.character.BigEnemy;
 import nl.yer.middlemen.dreampoint.character.Player;
 import nl.yer.middlemen.dreampoint.character.SmallEnemy;
 import nl.yer.middlemen.dreampoint.game.PiecesOnPlayingField;
-import nl.yer.middlemen.dreampoint.item.Item;
+import nl.yer.middlemen.dreampoint.item.Medkit;
 import nl.yer.middlemen.dreampoint.obstacle.Obstacle;
 import nl.yer.middlemen.dreampoint.obstacle.Tree;
-
-import java.util.Scanner;
 
 import static nl.yer.middlemen.dreampoint.game.Game.hiScore;
 
@@ -29,7 +27,7 @@ public class PlayingField {
 
         //Setting items, obstacles and random player position
         setObstacles(23, 55, 34, 77, 89);
-        setItems(22, 13, 50, 90, 64);
+        setMedkits(22, 13, 50, 90, 64);
         setEnemies(44, 69, 99);
         return map;
     }
@@ -47,16 +45,16 @@ public class PlayingField {
         }
     }
 
-    public void setItems(int first, int... rest){
-        Item item = new Item();
+    public void setMedkits(int first, int... rest){
+        Medkit medkit = new Medkit();
         int xPos = first % 10;
         int yPos = first / 10;
-        map[yPos][xPos] = item;
+        map[yPos][xPos] = medkit;
 
         for(int pos : rest){
             xPos = pos % 10;
             yPos = pos / 10;
-            map[yPos][xPos] = item;
+            map[yPos][xPos] = medkit;
         }
     }
 
@@ -98,9 +96,9 @@ public class PlayingField {
             System.out.println("Can't move through obstacle!");
             return true;
         }
-        else if (map[ypos][xpos] instanceof Item) {
+        else if (map[ypos][xpos] instanceof Medkit) {
             hiScore += 10;
-            System.out.println("Picked up an item!");
+            System.out.println("Picked up an medkit!");
             return false;
         }
         else if (map[ypos][xpos] instanceof SmallEnemy) {
